Added UnitSelector, state.unit string
UnitSelector is currently a copy of NameInput. State.unit is string from
that text field, but should be a value from a dropdown.

To support conversion without messing with user inputed values, each measure
now has two states that hold the value, one in the user's unit and the other
always in millimeters. The millimeters are passed to MaskPattern.

The conversion locations are marked with TODO comments.

Initial values have been moved to DEFAULT_MEASURES_IN_MM.